pyproject.toml
[tool.poetry]
name = "links-leeming-dev"
version = "0.1.0"
description = "Link page."
authors = ["Jonathan Leeming <jonathan@leeming.dev>"]
license = "MIT"
readme = "README.md"
packages = [{include = "links"}]

[tool.poetry.dependencies]
python = ">=3.11"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
